perf(test): load rebuild leaf tests natively
