+++ scripts/checks/test-create-require-budget.ts
@@ -23,11 +23,9 @@ export const CLI_CREATE_REQUIRE_FILES = [
   "src/lib/actions/sandbox/gateway-state-hints.test.ts",
   "src/lib/actions/sandbox/process-recovery-lock.test.ts",
   "src/lib/actions/sandbox/rebuild-agent-base-image-preflight.test.ts",
-  "src/lib/actions/sandbox/rebuild-config-hash.test.ts",
   "src/lib/actions/sandbox/rebuild-flow-helpers.test.ts",
   "src/lib/actions/sandbox/rebuild-gateway-drift.test.ts",
   "src/lib/actions/sandbox/rebuild-local-provider-recreate.test.ts",
-  "src/lib/actions/sandbox/rebuild-messaging-stage.test.ts",
   "src/lib/actions/sandbox/rebuild-resume-config.test.ts",
   "src/lib/actions/sandbox/rebuild-resume-reasoning.test.ts",
   "src/lib/actions/sandbox/sandbox-gateway-routing.test.ts",
